+++ src/reporter/tests/test_timescale_types.py
@@ -1,7 +1,6 @@
 import json
 import os
 import pytest
-import random
 import requests
 from requests import Response
 from typing import List, Union
@@ -14,6 +13,7 @@ from utils.kvt import merge_dicts
 
 from .geo_queries_fixture import query_1t1ena
 from .utils import notify_url
+from src.utils.tests.tenant import gen_tenant_id
 
 
 ENTITY_TYPE = 'device'
@@ -23,14 +23,6 @@ ENTITY_TYPE = 'device'
 # have to clean up the DB after each test, which would slow down the whole
 # test suite.
 #
-
-
-def gen_tenant_id() -> str:
-    tid = random.randint(1, 2**32)
-    return f"tenant{tid}"
-# TODO: duplicated code.
-# Move to some test util pkg that can be shared between reporter and
-# translators.
 
 
 TIMEX_ATTR_NAME = 'TimeInstant'