+++ RELEASE_NOTES.md
@@ -17,6 +17,7 @@
 - List Addix among adopters (#649)
 - Replaced string with the constants (#650)
 - Added idPattern ain query parameter (#648)
+- Remove duplicate code in src/reporter/tests/test_timescale_types.py (#657)
 
 ### Bug fixes